# Add note about `PRAGMA temp_store = memory` in sqlite setup