--- a/src/prefect/server/database/configurations.py
+++ b/src/prefect/server/database/configurations.py
@@ -421,7 +421,7 @@ class AioSqliteConfiguration(BaseDatabaseConfiguration):
         else:
             cursor.execute("PRAGMA busy_timeout = 60000;")  # 60s
 
-        # moves temporary tables from disk into RAM, speeds up queries
+        # moves temporary tables from disk into RAM, speeds up reads
         cursor.execute("PRAGMA temp_store = memory;")
 
         cursor.close()