--- a/src/prefect/server/database/configurations.py
+++ b/src/prefect/server/database/configurations.py
@@ -421,8 +421,10 @@ class AioSqliteConfiguration(BaseDatabaseConfiguration):
         else:
             cursor.execute("PRAGMA busy_timeout = 60000;")  # 60s
 
-        # moves temporary tables from disk into RAM, speeds up reads
-        cursor.execute("PRAGMA temp_store = memory;")
+        # `PRAGMA temp_store = memory;`moves temporary tables from disk into RAM
+        # this supposedly speeds up reads, but it seems to actually
+        # decrease overall performance, see https://github.com/PrefectHQ/prefect/pull/14812
+        # cursor.execute("PRAGMA temp_store = memory;")
 
         cursor.close()