related to #14658, we added `PRAGMA temp_store = memory;` to the setup in `AioSqliteConfiguration.setup_sqlite`

this moves temporary tables from disk into RAM, which (in theory) should improve read times

empirically, I am actually seeing a slight degradation in performance

so in this PR we just keep it commented out and add a note for the future to make sure we don't blindly add it without more exploration

---
### analysis

| Operation Type | Main Ops/s | Feature Ops/s | % Change |
|----------------|------------|---------------|----------|
| Simple Single  | 599.86     | 526.57        | -12.22%  |
| Simple Concurrent | 152.66  | 144.59        | -5.29%   |
| Complex Single | 29.70      | 25.49         | -14.18%  |
| Complex Concurrent | 7.22   | 6.03          | -16.48%  |

[benchmark used
](https://gist.github.com/zzstoatzz/4b2f6e4effc1ac4a77ad231fb2d04884)
notes:

1. Contrary to expectations, the `PRAGMA temp_store = memory` change this branch shows a performance decrease across all operation types.

2. The most significant performance drop is in complex concurrent operations, with a 16.48% decrease in operations per second.

3. Simple single-threaded operations, which were expected to benefit from in-memory storage, actually show a 12.22% decrease in performance.

4. Both branches show similar patterns in terms of concurrency impact: simple operations perform better in single-threaded mode, while complex operations have similar total execution times in both modes.

5. The performance degradation is consistent across all operation types and concurrency modes, suggesting that the `temp_store = memory` change may be having unintended consequences.